refact: Implement background job for gameworld creation (#85)
This PR introduces a `CreateGameworldJob` which runs the gameworld creation in a background process powered by `good_job`

Whenever a POST request is sent to `/gameworlds` we can now return the id of the gameworld instantly.
The gameworld creation is then done in the background.
As soon as the gameworld is completed we set it's status to `active` and publish the `gameworld-created` event to kafka

This PR also simplifies the `neighbours` as with our `x` and `y` coordinates we can simply query them from the database directly

This PR also introduce the pagination of the GET `/planets` endpoint. We now deliver 50 planets per page.